@@ -1,50 +1,51 @@
 {
-  "name": "react-chessground",
-  "version": "1.5.0",
-  "description": "React wrapper for chessground <https://github.com/ornicar/chessground>",
-  "main": "index.js",
-  "module": "es/index.js",
-  "scripts": {
-    "build": "gulp"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/ruilisi/react-chessground.git"
-  },
-  "keywords": [
-    "chess",
-    "chessboard",
-    "chessground",
-    "react"
-  ],
-  "authors": [
-    "Ruilisi"
-  ],
-  "license": "GPL-3.0",
-  "bugs": {
-    "url": "https://github.com/ruilisi/react-chessground/issues"
-  },
-  "homepage": "https://github.com/ruilisi/react-chessground",
+  "name": "react-chessground-example",
+  "version": "0.1.0",
+  "private": true,
   "dependencies": {
-    "chessground": "^7.6.12"
+    "@babel/preset-typescript": "^7.12.7",
+    "@types/chess.js": "^0.10.1",
+    "chess.js": "^0.10.3",
+    "chessground": "^7.9.3",
+    "eslint-config-prettier": "^4.3.0",
+    "html-webpack-plugin": "^3.2.0",
+    "less": "^3.13.1",
+    "react": "^16.14.0",
+    "react-chessground": "^1.5.0",
+    "react-dom": "^16.14.0",
+    "webpack-merge": "^4.2.2"
   },
-  "peerDependencies": {
-    "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1"
+  "scripts": {
+    "start": "webpack-dev-server --open --config webpack.dev.config.js",
+    "build": "webpack --config webpack.prod.config.js"
   },
   "devDependencies": {
-    "@babel/core": "^7.8.3",
-    "@babel/plugin-proposal-class-properties": "^7.8.3",
-    "@babel/preset-env": "^7.8.3",
-    "@babel/preset-react": "^7.8.3",
-    "@typescript-eslint/parser": "^4.14.0",
-    "eslint-import-resolver-typescript": "^2.3.0",
+    "@babel/core": "^7.12.10",
+    "@babel/plugin-proposal-class-properties": "^7.12.1",
+    "@babel/preset-env": "^7.12.11",
+    "@babel/preset-react": "^7.12.10",
+    "@types/react-dom": "^17.0.0",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-eslint": "^10.1.0",
+    "babel-loader": "^8.2.2",
+    "babel-plugin-import": "^1.13.3",
+    "css-loader": "^3.6.0",
+    "eslint": "5.3.0",
+    "eslint-config-airbnb": "17.1.0",
+    "eslint-loader": "^2.2.1",
+    "eslint-plugin-css-modules": "^2.11.0",
+    "eslint-plugin-html": "^5.0.5",
     "eslint-plugin-import": "^2.22.1",
-    "gulp": "^4.0.2",
-    "gulp-babel": "^8.0.0",
-    "gulp-concat": "^2.6.1",
-    "gulp-uglifycss": "^1.1.0",
-    "react-scripts": "^3.4.3"
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-prettier": "^3.3.1",
+    "eslint-plugin-react": "^7.22.0",
+    "file-loader": "^4.0.0",
+    "style-loader": "^0.23.1",
+    "ts-loader": "^8.0.14",
+    "typescript": "^4.1.3",
+    "webpack": "^4.46.0",
+    "webpack-cli": "^3.3.12",
+    "webpack-dev-server": "^3.11.2"
   }
 }
